Only look at the base path in readiness check
This is to ensure the readiness check for service containers, which depends on a file notification, works on Windows. We were comparing the event name with the path, but this path wasn't localized, so the match would fail on Windows. Instead, we can just look at the base of the path since we only set up the watcher on the containing directory.